Create ShubhanshuJha.java
This is the linear search implementation in Java.
In this program, if the key element is found in an array, the output will be "Yes", otherwise "No".
Moreover, the program is being implemented in the most efficient way a Linear Search can be done, i.e. in the best case (key == arr[0]), it will take O(1) time.